requirements.txt
Flask
cryptography
python-barcode
Pillow
gunicorn